#95 создал класс конфигурации для инициализации команд телеграм-бота внедрил мапу с командами в класс BotService, переименовал классы реализующие команды, перенёс всю логику из класса CommandHandler в класс Recommend, удалил класс CommandHandler, добавил метод форматирующий имя и фамилию пользователя в презентабельный вид
